+++ test/fixtures/parser/bigquery/create_function_no_args.sql
@@ -1,0 +1,3 @@
+CREATE FUNCTION add() RETURNS integer
+    AS 'select $1 + $2;'
+    LANGUAGE SQL;