+++ test/fixtures/parser/bigquery/create_table_options_as.sql
@@ -1,0 +1,3 @@
+CREATE OR REPLACE TABLE foo
+OPTIONS (description = 'copy of bar')
+AS (SELECT * from bar)