+++ test/fixtures/parser/bigquery/select_function_object_fields.sql
@@ -1,0 +1,8 @@
+SELECT
+  testFunction(a).b AS field,
+  testFunction(a).* AS wildcard,
+  testFunction(a).b.c AS field_with_field,
+  testFunction(a).b.* AS field_with_wildcard,
+  testFunction(a)[OFFSET(0)].* AS field_with_offset_wildcard,
+  testFunction(a)[ORDINAL(1)].* AS field_with_ordinal_wildcard
+FROM table1